#60 - feature(user specific views): introduce starting view
Currently, the home page was somewhat home and starting page at the same time, making the code and structure overall more confusing.

By introducing another page which contains the home content, the structure is more separated. The old home view is renamed to starting page, since it contains all other views and the navigation.

This signals that the first viw encountered is the starting page.