#60 - refactor(tests): restructure testing architecture
Due to the major restructuring to the firebase and user service, the whole testing structure does not work that well anymore.

Either the tests fail because certain functionality is not mocked properly or some files cant be accessed due to package limitations.

By creating a test_utils folder where both integration and widget tests can access, the latter on is solved.