#73 - feat(training plan): remove widget variable controller
As per recommendation of the getX maintainers, the controller should be injected in the first line of the build() method. This is in order to avoid issues with controllers not properly being disposed of when creating or deleting the widget.

If named routes are used, this issue should be fixed (will be implemented in a later change).