#73 - feature(training-plan): add unique key to editable table field
When swapping between different training plans the old state was sometimes kept due to flutter not being able to swap the displayed elements.

This resulted in a corrupted state which is now avoided, due to flutter being able to properly link the swapped elements and render objects with the widget.